+++ app/src/main/java/com/example/healthAppAndroid/historyTab/LiftingChart.java
@@ -27,7 +27,7 @@ public final class LiftingChart extends ChartContainer {
         setupChartView(xAxisFormatter);
     }
 
-    void updateChart(boolean isSmall, byte index) {
+    void updateChart(boolean isSmall, int index) {
         for (int i = 0; i < 4; ++i) {
             updateData(i, isSmall,
                        viewModel.entryRefs.get(index).get(i), i, viewModel.legendLabels[i]);
